requirements.txt
flask
requests
mysql-connector-python
gevent
